+++ src/roboto/version/autogen_version.py
@@ -1,4 +1,4 @@
-AUTOGEN_VERSION = "0.27.0"
+AUTOGEN_VERSION = "0.28.0"
 
 __all__= ("AUTOGEN_VERSION",)